+++ app/src/main/java/ru/euphoriadev/vk/api/model/VKResolveScreenName.java
@@ -11,13 +11,13 @@ public class VKResolveScreenName {
     public static final String TYPE_APP =   "application";
     public static final String TYPE_PAGE =  "page";
 
-    public long object_id;
+    public int object_id;
     public String type;
 
     public static VKResolveScreenName parse(JSONObject source) {
         VKResolveScreenName screenName = new VKResolveScreenName();
 
-        screenName.object_id = source.optLong("object_id");
+        screenName.object_id = source.optInt("object_id");
         screenName.type = source.optString("type");
         return screenName;
     }
